Extract PR-related models to their own module
